extract cellStatusPart
